Refactor ObjectKeyType, ObjectType, ObjectKeyView
