+++ cybsi/cloud/iocean/objects.py
@@ -6,10 +6,34 @@ from enum_tools import document_enum
 
 from ..enum import CybsiAPIEnum
 from ..internal import BaseAPI, BaseAsyncAPI, JsonObject, JsonObjectView
-from ..ioc import ObjectKeyType, ObjectKeyView, ObjectType
 from ..pagination import AsyncPage, Cursor, Page
 
 _PATH = "/iocean/collections/{}/objects"
+
+
+@document_enum
+class ObjectKeyType(CybsiAPIEnum):
+    """Object key type."""
+
+    MD5Hash = "MD5Hash"
+    SHA1Hash = "SHA1Hash"
+    SHA256Hash = "SHA256Hash"
+    SHA512Hash = "SHA512Hash"
+    DomainName = "DomainName"
+    URL = "URL"
+    IPAddress = "IPAddress"
+    IPNetwork = "IPNetwork"
+
+
+@document_enum
+class ObjectType(CybsiAPIEnum):
+    """Object type."""
+
+    File = "File"
+    DomainName = "DomainName"
+    URL = "URL"
+    IPAddress = "IPAddress"
+    IPNetwork = "IPNetwork"
 
 
 @document_enum
@@ -351,6 +375,20 @@ def _extract_changes_cursor(resp: httpx.Response) -> Optional[Cursor]:
     query = parse_qs(parsed.query)
     cursor = query["cursor"]
     return cast(Optional[Cursor], cursor[0]) if cursor is not None else None
+
+
+class ObjectKeyView(JsonObjectView):
+    """Object key view"""
+
+    @property
+    def type(self) -> ObjectKeyType:
+        """Object key type"""
+        return ObjectKeyType.from_string(self._get("type"))
+
+    @property
+    def value(self) -> str:
+        """Key value."""
+        return self._get("value")
 
 
 class ObjectView(JsonObjectView):
